use int64, not isize

+++ sort.c
@@ -38,7 +38,7 @@
 #endif
 
 #if !defined(LENGTH)
-#define LENGTH(x) (isize)((sizeof(x) / sizeof(*x)))
+#define LENGTH(x) (int64)((sizeof(x) / sizeof(*x)))
 #endif
 
 #if !defined(SIZEOF)
